+++ app/src/main/java/de/passwordvault/view/dialogs/ConfigureLoginDialog.java
@@ -3,7 +3,6 @@ package de.passwordvault.view.dialogs;
 import android.annotation.SuppressLint;
 import android.app.AlertDialog;
 import android.app.Dialog;
-import android.net.Uri;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
@@ -12,36 +11,27 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.DialogFragment;
 import androidx.lifecycle.ViewModelProvider;
 import de.passwordvault.R;
-import de.passwordvault.viewmodel.dialogs.CreateBackupViewModel;
+import de.passwordvault.viewmodel.dialogs.ConfigureLoginViewModel;
 
 
 /**
- * Class implements a {@linkplain Dialog} which allows the user to enter a password to create an
- * encrypted backup. The class that calls this dialog must always implement the interface
- * {@link de.passwordvault.view.utils.DialogCallbackListener}.
+ * Class implements a dialog to configure the application login.
  *
  * @author  Christian-2003
  * @version 3.2.0
  */
-public class CreateBackupDialog extends DialogFragment {
+public class ConfigureLoginDialog extends DialogFragment {
 
     /**
-     * Field contains the key that needs to be used when passing a directory as argument.
-     */
-    public static final String KEY_DIRECTORY = "directory";
-
-    /**
-     * Field contains the key that needs to be used when passing a
-     * {@link de.passwordvault.view.utils.DialogCallbackListener} as
-     * argument.
+     * Field stores the key that must be used when passing the callback listener as argument.
      */
     public static final String KEY_CALLBACK_LISTENER = "callback_listener";
 
 
     /**
-     * Attribute stores the {@link CreateBackupViewModel} for this {@link CreateBackupDialog}.
+     * Attribute stores the {@linkplain androidx.lifecycle.ViewModel} for the dialog.
      */
-    private CreateBackupViewModel viewModel;
+    private ConfigureLoginViewModel viewModel;
 
     /**
      * Attribute stores the view for the dialog.
@@ -50,36 +40,7 @@ public class CreateBackupDialog extends DialogFragment {
 
 
     /**
-     * Method returns the URI of the directory, into which the backup shall be saved.
-     *
-     * @return  URI of the directory of the backup.
-     */
-    public Uri getDirectory() {
-        return viewModel.getDirectory();
-    }
-
-    /**
-     * Method returns the password that was entered by the user. This is {@code null} if the user
-     * did not enter any password.
-     *
-     * @return  Password entered by the user.
-     */
-    public String getPassword() {
-        return viewModel.getPassword();
-    }
-
-    /**
-     * Method returns the filename, that was entered by the user.
-     *
-     * @return  Filename entered by the user.
-     */
-    public String getFilename() {
-        return viewModel.getFilename();
-    }
-
-
-    /**
-     * Method is called whenever a {@link CreateBackupDialog} is created.
+     * Method is called whenever a {@link ConfigureLoginDialog} is created.
      *
      * @param savedInstanceState    Previously saved state of the instance.
      * @return                      Created dialog.
@@ -89,7 +50,7 @@ public class CreateBackupDialog extends DialogFragment {
     @Override
     public Dialog onCreateDialog(@Nullable Bundle savedInstanceState) throws ClassCastException {
         super.onCreateDialog(savedInstanceState);
-        viewModel = new ViewModelProvider(this).get(CreateBackupViewModel.class);
+        viewModel = new ViewModelProvider(this).get(ConfigureLoginViewModel.class);
 
         try {
             viewModel.processArguments(getArguments());
@@ -99,13 +60,12 @@ public class CreateBackupDialog extends DialogFragment {
             //must surely be the stupidest decision made in this project so far...
             throw new ClassCastException(e.getMessage());
         }
-        viewModel.setFilename(getString(R.string.settings_security_backup_file));
 
-        view = requireActivity().getLayoutInflater().inflate(R.layout.dialog_create_backup, null);
+        view = requireActivity().getLayoutInflater().inflate(R.layout.dialog_configure_login, null);
 
         AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-        builder.setTitle(getString(R.string.settings_security_backup));
-        builder.setView(viewModel.createView(view));
+        builder.setTitle(R.string.settings_security_password);
+        builder.setView(view);
 
         builder.setPositiveButton(R.string.button_ok, (dialog, id) -> {
             //Action implemented in onStart()-method!
@@ -119,7 +79,7 @@ public class CreateBackupDialog extends DialogFragment {
 
 
     /**
-     * Method is called whenever the {@link CreateBackupDialog} is started. The method configures the
+     * Method is called whenever the {@link ConfigureLoginDialog} is started. The method configures the
      * click listeners for the dialog buttons.
      */
     @Override
@@ -134,20 +94,19 @@ public class CreateBackupDialog extends DialogFragment {
         //Configure positive button:
         Button positiveButton = dialog.getButton(Dialog.BUTTON_POSITIVE);
         positiveButton.setOnClickListener(view -> {
-            if (!viewModel.processUserInput(CreateBackupDialog.this.view)) {
+            if (!viewModel.processUserInput(ConfigureLoginDialog.this.view)) {
                 //Some inputs are incorrect:
                 return;
             }
             dismiss();
-            viewModel.getCallbackListener().onPositiveCallback(CreateBackupDialog.this);
+            viewModel.getCallbackListener().onPositiveCallback(ConfigureLoginDialog.this);
         });
 
         //Configure negative button:
         Button negativeButton = dialog.getButton(Dialog.BUTTON_NEGATIVE);
         negativeButton.setOnClickListener(view -> {
             dismiss();
-            viewModel.getCallbackListener().onNegativeCallback(CreateBackupDialog.this);
+            viewModel.getCallbackListener().onNegativeCallback(ConfigureLoginDialog.this);
         });
     }
-
 }